Split connector materialization helpers
