+++ crates/tandem-server/src/app/state/automation/logic.rs
@@ -8,4 +8,6 @@ include!("logic_parts/part04.rs");
 include!("logic_parts/part06.rs");
 include!("logic_parts/part08.rs");
 include!("logic_parts/part09.rs");
+include!("logic_parts/connector_materialization.rs");
+include!("logic_parts/connector_row_filter.rs");
 include!("logic_parts/part05.rs");
