+++ crates/tandem-server/src/app/state/automation/tests.rs
@@ -4,3 +4,4 @@ include!("tests_parts/part03.rs");
 include!("tests_parts/part04.rs");
 include!("tests_parts/part05.rs");
 include!("tests_parts/part06.rs");
+include!("tests_parts/part07.rs");
